requirements.txt
Flask==2.3.3
python-telegram-bot==20.7
gspread==5.12.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
requests==2.31.0
python-dotenv==1.0.0
